Clarify comments' use of the word "support"
It was initially stated that "[PowerShell] has no shortcut support",
which is not *strictly* true as PowerShell can interact with COM objects
to interact with shortcuts objects.

PowerShell has similarly been able to interact with symlinks, hardlinks,
and junctions for a while;  the significance that the comment pertaining
to links now points out is that PowerShell has native link support
starting with 5.0.