Use the full source path when making a shortcut
The WScript.Shell COM object allows the creation of shortcut with any
path as an argument, but does not resolve paths during creation so the
argument must be a full (absolute) path or the resulting shortcut file
will be broken.  This commit provides the logic to work around that
restriction by resolving the full path of the `Source` parameter prior
to passing it to WScript.Shell.